requirements.txt
GitPython==3.1.24
python-gitlab==2.10.1
